Refactor password gen into FreeIPAHelpers
Refactors various bits of code to consolidate the helper methods into a
single class
